book/gen/pyproject.toml
[project]
name = "gen"
version = "0.1.0"
description = "Generates a book from the devnets repo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.5",
    "pyyaml>=6.0.2",
]
